+++ test/libsolidity/syntaxTests/storageLayoutSpecifier/duplicated_layout_definition.sol
@@ -1,3 +1,3 @@
 contract C layout at 0x1234 is A, B layout at 0xABC { }
 // ----
-// ParserError 8714: (36-42): Storage base location was already defined previously.
+// ParserError 8714: (36-42): Storage layout was already defined previously.